Migrate date-time single arg tests
