+++ datetimeformatters/src/androidTest/java/dev/drewhamilton/androidtime/format/AndroidDateTimeFormatterTest.kt
@@ -19,132 +19,6 @@ import java.util.Locale
 @RequiresApi(21) // Instrumented tests for Dynamic Features is not supported on API < 21 (AGP 4.0.0-beta04)
 class AndroidDateTimeFormatterTest : TimeSettingTest() {
 
-    //region ofLocalizedDateTime with dateTimeStyle
-    @Test fun ofLocalizedDateTime_nullSystemSettingUsLocaleShortDateTimeFormat_usesShort12HourUsFormat() {
-        assumeNullableSystemTimeSetting()
-
-        systemTimeSetting = null
-        testLocale = Locale.US
-
-        val formatter = AndroidDateTimeFormatter.ofLocalizedDateTime(testContext, FormatStyle.SHORT)
-
-        val result = formatter.format(DATE_TIME)
-        assertThat(result).contains("4/24/10")
-        assertThat(result).contains("4:44 PM")
-    }
-
-    @Test fun ofLocalizedDateTime_12SystemSettingUsLocaleShortDateTimeFormat_usesShort12HourUsFormat() {
-        systemTimeSetting = TIME_SETTING_12
-        testLocale = Locale.US
-
-        val formatter = AndroidDateTimeFormatter.ofLocalizedDateTime(testContext, FormatStyle.SHORT)
-
-        val result = formatter.format(DATE_TIME)
-        assertThat(result).contains("4/24/10")
-        assertThat(result).contains("4:44 PM")
-    }
-
-    @Test fun ofLocalizedDateTime_24SystemSettingUsLocaleShortDateTimeFormat_usesShort24HourUsFormat() {
-        systemTimeSetting = TIME_SETTING_24
-        testLocale = Locale.US
-
-        val formatter = AndroidDateTimeFormatter.ofLocalizedDateTime(testContext, FormatStyle.SHORT)
-
-        val result = formatter.format(DATE_TIME)
-        assertThat(result).contains("4/24/10")
-        assertThat(result).contains("16:44")
-    }
-
-    @Test fun ofLocalizedDateTime_usLocaleMediumDateTimeFormat_usesMediumUsFormat() {
-        testLocale = Locale.US
-
-        val formatter = AndroidDateTimeFormatter.ofLocalizedDateTime(testContext, FormatStyle.MEDIUM)
-
-        val result = formatter.format(DATE_TIME)
-        assertThat(result).contains("Apr 24, 2010")
-        assertThat(result).contains("4:44:00 PM")
-    }
-
-    @Test fun ofLocalizedDateTime_usLocaleLongDateTimeFormat_usesLongUsFormat() {
-        testLocale = Locale.US
-
-        val formatter = AndroidDateTimeFormatter.ofLocalizedDateTime(testContext, FormatStyle.LONG)
-
-        val result = formatter.format(DATE_TIME)
-        assertThat(result).contains("April 24, 2010")
-        assertThat(result).contains("4:44:00 PM Z")
-    }
-
-    @Test fun ofLocalizedDateTime_usLocaleFullDateTimeFormat_usesFullUsFormat() {
-        testLocale = Locale.US
-
-        val formatter = AndroidDateTimeFormatter.ofLocalizedDateTime(testContext, FormatStyle.FULL)
-
-        val result = formatter.format(DATE_TIME)
-        assertThat(result).contains("Saturday, April 24, 2010")
-        assertThat(result).contains("4:44:00 PM Z")
-    }
-
-    @Test fun ofLocalizedDateTime_nullSystemSettingItalyLocaleShortDateTimeFormat_usesShort24HourItalyFormat() {
-        assumeNullableSystemTimeSetting()
-
-        systemTimeSetting = null
-        testLocale = Locale.ITALY
-
-        val formatter = AndroidDateTimeFormatter.ofLocalizedDateTime(testContext, FormatStyle.SHORT)
-
-        val result = formatter.format(DATE_TIME)
-        assertThat(result).contains("24/04/10")
-        assertThat(result).contains("16:44")
-    }
-
-    @Test fun ofLocalizedDateTime_12SystemSettingItalyLocaleShortDateTimeFormat_usesShort12HourItalyFormat() {
-        systemTimeSetting = TIME_SETTING_12
-        testLocale = Locale.ITALY
-
-        val formatter = AndroidDateTimeFormatter.ofLocalizedDateTime(testContext, FormatStyle.SHORT)
-
-        val result = formatter.format(DATE_TIME)
-        assertThat(result).contains("24/04/10")
-        assertThat(result).contains("4:44 PM")
-    }
-
-    @Test fun ofLocalizedDateTime_24SystemSettingItalyLocaleShortDateTimeFormat_usesShort24HourItalyFormat() {
-        systemTimeSetting = TIME_SETTING_24
-        testLocale = Locale.ITALY
-
-        val formatter = AndroidDateTimeFormatter.ofLocalizedDateTime(testContext, FormatStyle.SHORT)
-
-        val result = formatter.format(DATE_TIME)
-        assertThat(result).contains("24/04/10")
-        assertThat(result).contains("16:44")
-    }
-
-    @Test fun ofLocalizedDateTime_italyLocaleMediumDateTimeFormat_usesMediumItalyFormat() {
-        testLocale = Locale.ITALY
-
-        val formatter = AndroidDateTimeFormatter.ofLocalizedDateTime(testContext, FormatStyle.MEDIUM)
-
-        val result = formatter.format(DATE_TIME)
-        assertThat(result).contains(ITALY_MEDIUM_DATE)
-        assertThat(result).contains(ITALY_MEDIUM_TIME)
-    }
-
-    @Test fun ofLocalizedDateTime_italyLocaleLongDateTimeFormat_usesLongItalyFormat() {
-        testLocale = Locale.ITALY
-
-        val formatter = AndroidDateTimeFormatter.ofLocalizedDateTime(testContext, FormatStyle.LONG)
-        assertEquals("24 aprile 2010 16:44:00 Z", formatter.format(DATE_TIME))
-    }
-
-    @Test fun ofLocalizedDateTime_italyLocaleFullDateTimeFormat_usesFullItalyFormat() {
-        testLocale = Locale.ITALY
-
-        val formatter = AndroidDateTimeFormatter.ofLocalizedDateTime(testContext, FormatStyle.FULL)
-        assertEquals("sabato 24 aprile 2010 16:44:00 Z", formatter.format(DATE_TIME))
-    }
-    //endregion
-
     //region ofLocalizedDateTime with dateStyle and timeStyle
     @Test fun ofLocalizedDateTime_usLocaleShortDateFormatLongTimeFormat_usesShortDateLongTimeUsFormat() {
         testLocale = Locale.US
